+++ src/limits.h
@@ -4,11 +4,11 @@
 #include <iostream>
 
 class Limits {
-    double _min;
-    double _max;
+    float _min;
+    float _max;
 public:
-    Limits() : _min(std::numeric_limits<double>::max()), _max(std::numeric_limits<double>::lowest()) {}
-    Limits(const double min, const double max) : _min(min), _max(max) {}
-    void update(const double value);
+    Limits() : _min(std::numeric_limits<float>::max()), _max(std::numeric_limits<float>::lowest()) {}
+    Limits(const float min, const float max) : _min(min), _max(max) {}
+    void update(const float value);
     friend std::ostream& operator << (std::ostream& out, const Limits &limits);
 };
