+++ src/point.cpp
@@ -2,21 +2,21 @@
 
 #include "point.h"
 
-point::point(const double x, const double y, const double z) {
+point::point(const float x, const float y, const float z) {
     R[0] = x;
     R[1] = y;
     R[2] = z;
 }
 
-double point::length() const {
-    return sqrt(R[0] * R[0] + R[1] * R[1] + R[2] * R[2]);
+float point::length() const {
+    return std::sqrt(R[0] * R[0] + R[1] * R[1] + R[2] * R[2]);
 }
 
-double point::length_squared() const {
+float point::length_squared() const {
     return R[0] * R[0] + R[1] * R[1] + R[2] * R[2];
 }
 
-double point::radius_squared() const {
+float point::radius_squared() const {
     return R[0] * R[0] + R[1] * R[1];
 }
 
@@ -36,11 +36,11 @@ point operator - (const point& a, const point& b) {
     return c;
 }
 
-double operator * (const point& a, const point& b) {
+float operator * (const point& a, const point& b) {
     return a.R[0] * b.R[0] + a.R[1] * b.R[1] + a.R[2] * b.R[2];
 }
 
-point operator * (const point& a, const double& b) {
+point operator * (const point& a, const float& b) {
     point c;
     for (int i = 0; i < 3; ++i) {
         c.R[i] = a.R[i] * b;
@@ -48,7 +48,7 @@ point operator * (const point& a, const double& b) {
     return c;
 }
 
-point operator * (const double& a, const point& b) {
+point operator * (const float& a, const point& b) {
     point c;
     for (int i = 0; i < 3; ++i) {
         c.R[i] = a * b.R[i];
@@ -56,7 +56,7 @@ point operator * (const double& a, const point& b) {
     return c;
 }
 
-point operator / (const point& a, const double& b) {
+point operator / (const point& a, const float& b) {
     point c;
     for (int i = 0; i < 3; ++i) {
         c.R[i] = a.R[i] / b;
